fluxdiff/requirements.txt
opencv-python
numpy
pillow
click
flask
cairosvg
PyYAML